Extract ParBuilder from Interpreter
