+++ casper/src/test/scala/coop/rchain/casper/util/rholang/RuntimeManagerTest.scala
@@ -13,7 +13,7 @@ import coop.rchain.metrics
 import coop.rchain.metrics.Metrics
 import coop.rchain.p2p.EffectsTestInstances.LogicalTime
 import coop.rchain.rholang.Resources.mkRuntime
-import coop.rchain.rholang.interpreter.{accounting, Interpreter}
+import coop.rchain.rholang.interpreter.{accounting, Interpreter, ParBuilder}
 import coop.rchain.rholang.interpreter.accounting.Cost
 import coop.rchain.shared.Log
 import monix.eval.Task
@@ -65,7 +65,7 @@ class RuntimeManagerTest extends FlatSpec with Matchers {
                           val initialPhlo = Cost(accounting.MAX_VALUE)
                           for {
                             _             <- runtime.reducer.setPhlo(initialPhlo)
-                            term          <- Interpreter[Task].buildNormalizedTerm(deploy.term)
+                            term          <- ParBuilder[Task].buildNormalizedTerm(deploy.term)
                             _             <- runtime.reducer.inj(term)
                             phlosLeft     <- runtime.reducer.phlo
                             reductionCost = initialPhlo - phlosLeft
