+++ rholang/src/test/scala/coop/rchain/rholang/interpreter/NormalizeTest.scala
@@ -59,7 +59,7 @@ class CollectMatcherSpec extends FlatSpec with Matchers {
     IndexMapChain[VarSort]().newBindings(List(("P", ProcSort, 0, 0), ("x", NameSort, 0, 0))),
     DebruijnLevelMap[VarSort]()
   )
-  def getNormalizedPar(rho: String): Par = Interpreter[Coeval].buildNormalizedTerm(rho).value()
+  def getNormalizedPar(rho: String): Par = ParBuilder[Coeval].buildNormalizedTerm(rho).value()
   def assertEqualNormalized(rho1: String, rho2: String): Assertion =
     assert(getNormalizedPar(rho1) == getNormalizedPar(rho2))
 
@@ -426,45 +426,45 @@ class ProcMatcherSpec extends FlatSpec with Matchers {
 
   "PSend" should "Not compile if data contains negation" in {
     an[TopLevelLogicalConnectivesNotAllowedError] should be thrownBy {
-      Interpreter[Coeval].buildNormalizedTerm("""new x in { x!(~1) }""").value()
+      ParBuilder[Coeval].buildNormalizedTerm("""new x in { x!(~1) }""").value()
     }
   }
 
   "PSend" should "Not compile if data contains conjunction" in {
     an[TopLevelLogicalConnectivesNotAllowedError] should be thrownBy {
-      Interpreter[Coeval].buildNormalizedTerm("""new x in { x!(1 /\ 2) }""").value()
+      ParBuilder[Coeval].buildNormalizedTerm("""new x in { x!(1 /\ 2) }""").value()
     }
   }
 
   "PSend" should "Not compile if data contains disjunction" in {
     an[TopLevelLogicalConnectivesNotAllowedError] should be thrownBy {
-      Interpreter[Coeval].buildNormalizedTerm("""new x in { x!(1 \/ 2) }""").value()
+      ParBuilder[Coeval].buildNormalizedTerm("""new x in { x!(1 \/ 2) }""").value()
     }
   }
 
   "PSend" should "Not compile if data contains wildcard" in {
     an[TopLevelWildcardsNotAllowedError] should be thrownBy {
-      Interpreter[Coeval].buildNormalizedTerm("""@"x"!(_)""").value()
+      ParBuilder[Coeval].buildNormalizedTerm("""@"x"!(_)""").value()
     }
   }
 
   "PSend" should "Not compile if data contains free variable" in {
     an[TopLevelFreeVariablesNotAllowedError] should be thrownBy {
-      Interpreter[Coeval].buildNormalizedTerm("""@"x"!(y)""").value()
+      ParBuilder[Coeval].buildNormalizedTerm("""@"x"!(y)""").value()
     }
   }
 
   "PSend" should "not compile if name contains connectives" in {
     an[TopLevelLogicalConnectivesNotAllowedError] should be thrownBy {
-      Interpreter[Coeval].buildNormalizedTerm("""@{Nil /\ Nil}!(1)""").value()
+      ParBuilder[Coeval].buildNormalizedTerm("""@{Nil /\ Nil}!(1)""").value()
     }
 
     an[TopLevelLogicalConnectivesNotAllowedError] should be thrownBy {
-      Interpreter[Coeval].buildNormalizedTerm("""@{Nil \/ Nil}!(1)""").value()
+      ParBuilder[Coeval].buildNormalizedTerm("""@{Nil \/ Nil}!(1)""").value()
     }
 
     an[TopLevelLogicalConnectivesNotAllowedError] should be thrownBy {
-      Interpreter[Coeval].buildNormalizedTerm("""@{~Nil}!(1)""").value()
+      ParBuilder[Coeval].buildNormalizedTerm("""@{~Nil}!(1)""").value()
     }
   }
 
@@ -784,45 +784,45 @@ class ProcMatcherSpec extends FlatSpec with Matchers {
 
   "PInput" should "not compile when connectives are used in the channel" in {
     an[TopLevelLogicalConnectivesNotAllowedError] should be thrownBy {
-      Interpreter[Coeval]
+      ParBuilder[Coeval]
         .buildNormalizedTerm("""for(x <- @{Nil \/ Nil}){ Nil }""")
         .value()
     }
 
     an[TopLevelLogicalConnectivesNotAllowedError] should be thrownBy {
-      Interpreter[Coeval]
+      ParBuilder[Coeval]
         .buildNormalizedTerm("""for(x <- @{Nil /\ Nil}){ Nil }""")
         .value()
     }
 
     an[TopLevelLogicalConnectivesNotAllowedError] should be thrownBy {
-      Interpreter[Coeval].buildNormalizedTerm("""for(x <- @{~Nil}){ Nil }""").value()
+      ParBuilder[Coeval].buildNormalizedTerm("""for(x <- @{~Nil}){ Nil }""").value()
     }
   }
 
   "PInput" should "not compile when connectives are the top level expression in the body" in {
     an[TopLevelLogicalConnectivesNotAllowedError] should be thrownBy {
-      Interpreter[Coeval].buildNormalizedTerm("""for(x <- @Nil){ 1 /\ 2 }""").value()
+      ParBuilder[Coeval].buildNormalizedTerm("""for(x <- @Nil){ 1 /\ 2 }""").value()
     }
 
     an[TopLevelLogicalConnectivesNotAllowedError] should be thrownBy {
-      Interpreter[Coeval].buildNormalizedTerm("""for(x <- @Nil){ 1 \/ 2 }""").value()
+      ParBuilder[Coeval].buildNormalizedTerm("""for(x <- @Nil){ 1 \/ 2 }""").value()
     }
 
     an[TopLevelLogicalConnectivesNotAllowedError] should be thrownBy {
-      Interpreter[Coeval].buildNormalizedTerm("""for(x <- @Nil){ ~1 }""").value()
+      ParBuilder[Coeval].buildNormalizedTerm("""for(x <- @Nil){ ~1 }""").value()
     }
   }
 
   "PInput" should "not compile when logical OR or NOT is used in the pattern of the receive" in {
     an[PatternReceiveError] should be thrownBy {
-      Interpreter[Coeval]
+      ParBuilder[Coeval]
         .buildNormalizedTerm("""new x in { for(@{Nil \/ Nil} <- x) { Nil } }""")
         .value()
     }
 
     an[PatternReceiveError] should be thrownBy {
-      Interpreter[Coeval]
+      ParBuilder[Coeval]
         .buildNormalizedTerm("""new x in { for(@{~Nil} <- x) { Nil } }""")
         .value()
     }
@@ -830,7 +830,7 @@ class ProcMatcherSpec extends FlatSpec with Matchers {
 
   "PInput" should "compile when logical AND is used in the pattern of the receive" in {
     noException should be thrownBy {
-      Interpreter[Coeval]
+      ParBuilder[Coeval]
         .buildNormalizedTerm("""new x in { for(@{Nil /\ Nil} <- x) { Nil } }""")
         .value()
     }
@@ -838,7 +838,7 @@ class ProcMatcherSpec extends FlatSpec with Matchers {
 
   "PContr" should "not compile when logical OR or NOT is used in the pattern of the receive" in {
     an[PatternReceiveError] should be thrownBy {
-      Interpreter[Coeval]
+      ParBuilder[Coeval]
         .buildNormalizedTerm(
           new StringReader("""new x in { contract x(@{ y /\ {Nil \/ Nil}}) = { Nil } }""")
         )
@@ -846,7 +846,7 @@ class ProcMatcherSpec extends FlatSpec with Matchers {
     }
 
     an[PatternReceiveError] should be thrownBy {
-      Interpreter[Coeval]
+      ParBuilder[Coeval]
         .buildNormalizedTerm(
           new StringReader("""new x in { contract x(@{ y /\ ~Nil}) = { Nil } }""")
         )
@@ -856,7 +856,7 @@ class ProcMatcherSpec extends FlatSpec with Matchers {
 
   "PContr" should "compile when logical AND is used in the pattern of the receive" in {
     noException should be thrownBy {
-      Interpreter[Coeval]
+      ParBuilder[Coeval]
         .buildNormalizedTerm(
           new StringReader("""new x in { contract x(@{ y /\ {Nil /\ Nil}}) = { Nil } }""")
         )
@@ -1518,7 +1518,7 @@ class ProcMatcherSpec extends FlatSpec with Matchers {
            }
          }
        """
-        Interpreter[Coeval].buildNormalizedTerm(rho).value()
+        ParBuilder[Coeval].buildNormalizedTerm(rho).value()
         assert(true)
       } catch {
         case e: Throwable =>
