+++ rholang/src/test/scala/coop/rchain/rholang/interpreter/PrettyPrinterTest.scala
@@ -834,7 +834,7 @@ class ProcPrinterSpec extends FlatSpec with Matchers {
     assert(parseAndPrint(prettySource) == prettySource)
 
   private def parseAndPrint(source: String): String = PrettyPrinter().buildString(
-    Interpreter[Coeval].buildNormalizedTerm(new StringReader(source)).runAttempt().right.get
+    ParBuilder[Coeval].buildNormalizedTerm(new StringReader(source)).runAttempt().right.get
   )
 }
