+++ rholang/src/test/scala/coop/rchain/rholang/interpreter/accounting/CostAccountingPropertyTest.scala
@@ -39,7 +39,7 @@ class CostAccountingPropertyTest extends FlatSpec with PropertyChecks with Match
 
   implicit val taskExecutionDuration: FiniteDuration = 5.seconds
 
-  def cost(proc: Proc): Cost = Cost(Interpreter[Coeval].buildPar(proc).apply)
+  def cost(proc: Proc): Cost = Cost(ParBuilder[Coeval].buildPar(proc).apply)
 
   behavior of "Cost accounting in Reducer"
 
@@ -87,15 +87,12 @@ object CostAccountingPropertyTest {
       .map { _.sliding(2).forall { case List(r1, r2) => r1 == r2 } }
       .runSyncUnsafe(duration)
 
-  def execute[F[_]: Sync](runtime: Runtime[F], p: Proc): F[Long] = {
-    val interpreter = Interpreter[F]
-
+  def execute[F[_]: Sync](runtime: Runtime[F], p: Proc): F[Long] =
     for {
-      program <- interpreter.buildPar(p)
+      program <- ParBuilder[F].buildPar(p)
       res     <- evaluatePar(runtime, program)
       cost    = res.cost
     } yield cost.value
-  }
 
   def evaluatePar[F[_]: Sync](
       runtime: Runtime[F],
